Change: Use compose spec based compose file for better startup
It seems under strange circumstances which are reproducible on my
machine ospd-openvas doesn't load the VT data due to a race condition
with the vulnerabilities-test container startup.

To fix this issue require docker-compose 1.27.0 or newer that implements
the new compose spec which provides conditions for the startup
procedures. This allows the ospd-openvas container to wait until the
vulnerabilities-tests container has finished its job.