Change: Don't use `--prefix` argument with pip
Using the `--prefix` argument is broken with newer pip versions and at
least Python 3.10. The distros are patching CPython's sysconfig module
which breaks our expected behavior if the `--prefix` argument . Because
the default install prefix is already `/usr/local` we just can remove
the `--prefix` argument to circumvent the bug.

See the following links for details:

* https://github.com/pypa/pip/issues/11651
* https://github.com/pypa/pip/issues/10978
* https://discuss.python.org/t/linux-distro-patches-to-sysconfig-are-changing-pip-install-prefix-outside-virtual-environments/18240