Improve python software installations (#265)
* Change: Improve installing Python applications

Use a venv to prepare the to be installed files. Within a virtual
environment the `--prefix` and `--root` arguments of pip work as
expected and aren't influenced by the distribution specific patches to
the system wide installed Python version.

* Change: Let system Python find our installed applications and packages

Sadly with using a virtual environment the installation directory is
named site-packages. This site-packages directory is not in the Python
module path (aka. PYTHONPATH, sys.path). Therefore add a .pth file to
add the directory to the system Python module search path.

---------

Co-authored-by: Jaspar S <jaspar.stach@greenbone.net>