Fix: Avoid permission errors when setting up postgres db
Depending on the default settings for creating a user the current
working directory may not be executable and readable for the postgres
user. When running the commands for setting up the postgres database
in this home directory permission errors may be raised. To avoid that
change into the home directory of the postgres user before setting up
the database.