Fix: Quote variables for well defined behavior By quoting variables we can improve stability of this script. $HOME and $password are both user supplied variables, that may contain spaces.
We need to quote them, or they may be expanded into extra arguments.
This means this scripit will no longer break, if a space is supplied
inside the password.

`read` treats backslashes differently, which can be annoying when
entering a random password. with the `-r` flag we can disable this
behavior.

`$@` will return an array of all arguments, which will then be
translated into a string. When using `$*` we will get a sting which is
safe to use, as its behavior is defined.